Added a cylendar around player which manages health
