+++ Assets/Materials/HealthCylinder.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0267b5b5b6846d04aaa315a73baf5d81
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
