+++ Assets/Scripts/PawnScripts/PlayerHealthRegen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f76705d67a452b40a15c7b109e81f22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
